#83 [MOD] 서비스 분리 & 북마크 연동
